Moved tests from MessageParser to ChatManager test class
